feat: move getSuperJson among prepare modules
